src/main/java/pl/tomek/controller/LoginController.java
package pl.tomek.controller;

import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseBody;

@Controller
public class LoginController {


    @GetMapping("/loginform")
    public String loginForm(Model model) {


        return "login_form";
    }

    @GetMapping("/formError")
    public String errors()
    {
        return "formErrors";
    }





}
